Add an internal 'bytes' package
This package is designed to read/write uint8, uint16, uint32, and uint64
values from a byte array. The 'Bytes' interface allows you to access the
underlying byte array as either BigEndian or LittleEndian as well as
slice into smaller regions of the larger byte array without any copying.
This package forms the foundation of an 'mmio' library built on top of
it to perform read/write operations on mmaped files or memory.

Signed-off-by: Kevin Klues <kklues@nvidia.com>