Add an internal 'nvml' package
This package is a wrapper around the upstream 'go-nvml' package,
providing a mockable interface around the NVML calls required by the
other packages in this project. As part of this, it provides a mock
implementation of a Luna server with eight A100 devices on it. This
mocked Luna server provides the basis for other packages in this project
to implement their unit tests on top of.

Signed-off-by: Kevin Klues <kklues@nvidia.com>