Add an 'nvpci' package
This package is designed to interact with a GPU directly over PCIe (as
opposed to through the NVIDIA driver). It works with the files under
/sys/bus/pci/devices/* and provids an abstraction to both enumerate GPUs
as well as read/write data to/from a GPUs PCIe config-space and memory
mapped BAR regions. Convenience functions exist for extracting things
like the vendor, device id, class id, etc.

Signed-off-by: Kevin Klues <kklues@nvidia.com>