Add deployment for nvidia-mig-manager.service through systemd
The current implementation has been built on a DGX-A100 system (running
the latest DGX OS -- a variant of Ubuntu 20.04). It is therefore
customized to the set of services and components expected to be running
on this OS. The file 'utils-custom.sh' can be used to customize this for
other deployments without changing the core logic of the service and its
constituent components.

Signed-off-by: Kevin Klues <kklues@nvidia.com>