Update GPU reset logic to only reset GPUs that require it
In the case of *no* NVIDIA module loaded, this means only reset those
GPUs that have a MIG mode change pending. When the NVIDIA module *is*
loaded, however, we still need to reset (almost) all of the GPUs because
any of them connected by an NVLINK must be reset together or else the
reset fails. We can be a little smarter though, and only reset any GPUs
that are "3D controllers", filtering out any "VGA Controllers". This is,
in fact, required on something like the DGX station, where there are 4
A100s and a single VGA Controller attached to the machine.

Signed-off-by: Kevin Klues <kklues@nvidia.com>