Support strings or arrays as input to v1.Spec.DeviceFilter
As part of this, I moved the "Normalize()" function for
MigConfigSpecSlice into the export package since it is actually very
specific to this command. I also renamed this command to "merge" instead
of "normalize" and added a bunch of tests for it.

This function converts a flat slice of the MIG config for each device
individually, and normalizes it to merge like-configured devices, change
the Devices list to "all" when appropriate, and update the DeviceFilter
as appropriate. This makes for a better display when printing the output
from the 'export' command.

Signed-off-by: Kevin Klues <kklues@nvidia.com>