Implementation of a 'mig-manager' for the GPU Operator
This MIG manager ss designed to run as a daemonset and will respond to
changes to the label 'nvidia.com/mig.config', triggering the node to be
reconfigured with the 'nvidia-mig-parted' configuration specified in
this label.

The example 'nvidia-mig-manager-example.yaml' can be deployed on a
cluster with the GPU Operator running to install the manager and test
it, independent of the one installed by the operator itself.

Signed-off-by: Kevin Klues <kklues@nvidia.com>