Add CI for container images
This change adds basic CI for the container images generated from
deployments/gpu-operator for the mig-manager

The CI is reworked to use predefined build, push, and pull dummy steps
to construct the desired actions for the mig-manager container images.
These are combined with dummy steps that define the input and output image
and registry.

The internal CI is modified to pull a sha-tagged image from the public
repository, run the scans and release this, instead of repeating the
build process.

Signed-off-by: Evan Lezar <elezar@nvidia.com>