Add nvidia-dcgm.service to driver_services array
Beginning with DCGM 2.1, the datacenter-gpu-manager package
now provides both a dcgm.service and an nvidia-dcgm.service.
They conflict with one another so that only one can run at a
time. I'm not sure why they didn't just use an alias, but
update the driver_services array so that we can handle both
cases:

 /lib/systemd/system/nvidia-dcgm.service:
    [Unit]
    Description=NVIDIA DCGM service
    Conflicts=dcgm.service
    ...

 /lib/systemd/system/dcgm.service:
    [Unit]
    Description=DEPRECATED. Please use nvidia-dcgm.service
    Conflicts=nvidia-dcgm.service
    ...