Remove CombinedMigManager and add wrappers for Mode/Config Managers
These new wrappers encapsulate the logic to decide if a PCI based
manager or an NVML based manager should be instantiated based on (1)
availability of the NVIDIA driver and (2) a minimum NVML version
requirement.

As part of this, the logic that uses these managers was updated to
ensure that each type of manager is only instantiated if/when it is
needed. For example, if an assert is happening and we have already
determined that a GPU does not have MIG mode enabled, we will not
instantiate a MIG Config manager to then assert the config. This allows
us to continue running on old hardware without MIG support and / or
older versions of NVML that don't have the supported MIG calls in them.

A large part of this refactoring was motivated by a bug where enabling
the CONFIG_IO_STRICT_DEVMEM kernel flag prevents us from reading from BAR0
memory if a driver is already attached to it. This patch avoids
this issue by making sure the NVML based version of the MIG Mode Manager
is used, so long as it is available (which uses the driver under the
hood instead of trying to read the BAR0 memory directly).

Signed-off-by: Kevin Klues <kklues@nvidia.com>