Keep NVML alive across all mig-parted commands (except GPU reset)
A number of bugs were reported that the nvidia-mig-manager was delaying boot
times of DGX OS by up to 40 minutes! This doesn't happen with v0.1.2 of
mig-parted (which is what currently ships with DGX OS), but started happening
with v0.4.0.

Interestingly, this doesn't happen with v0.4.0 if you first put the GPUs into
persistence mode.

After further investigation, I found that the reason this is happening is that
each individual operation in mig-parted connects to the nvidia driver via a
call to nvml.Init(), followed by an nvml.Shutdown() when it is finished. That
is, we run this Init/Shutdown combination for ever single mig-capable check,
every single mig-mode check, every single mig-mode switch, etc.

Since the persistence daemon is not yet started when the nvidia-mig-parted
service runs, every one of these Init/Shutdown combinations fully initializes
the driver and then fully shuts it down. This operation can take up to a minute
from a sleeping driver, and every operation we do is triggering it to go back
to sleep before the next operation wakes it up. This adds up for each
operation, making it very slow to complete any individual mig-parted call.

Luckily the fix is fairly simple – init the driver once and keep it awake
across the entirety of a mig-parted call. The one caveat being that we need to
perform a GPU reset when switching MIG mode, so we have to release it and
reaquire it across this reset call.

The reason this didn't happen on v0.1.2 is that we didn't actually rely on NVML
to do our MIG mode checks (or switching). We used the PCI based
MigModeManager() instead. We still had the same issue with switching MIG
configs, but since most of the operations were not subject to this issue, it
didn't take quite as long and went unnoticed.

Signed-off-by: Kevin Klues <kklues@nvidia.com>