Add extra assert in k8s-mig-manager to double check mig-mode change applied
On some systems 'nvidia-smi -r'  will succeed even if the underlying GPU reset
wasn't actually carried out. This double checks that the intended mode change
was actually applied after the 'apply' command returns 'SUCCESS'. In the future
we should consider embedding this double check into the 'apply'  call itself so
that it would nevr return 'SUCCESS' in the first place.

Signed-off-by: Kevin Klues <kklues@nvidia.com>