packaging: Fix postinst failure when nvidia-smi is missing
If the the nvidia-smi binary is not available on the system, then
the following call will always fail:

  which nvidia-smi > /dev/null 2>&1

This causes the postinst script to error out, leaving the package
in a half-configure state. Update the nvidia-smi check so that the
postinst script is able to tolerate this binary not being present.
Also create the "hooks-default" symlink when this happens.

Lastly, update the .spec file with the same change.

Signed-off-by: Jamie Nguyen <jamien@nvidia.com>