Update all calls to enumerate GPUs to use NVML or PCI as appropriate
Previously, all GPU enumeration was done by looking directly at the PCI bus. If
any GPUs had become unusable by the driver (and were not enumerable by NVML
anymore), this would cause problems (because we would try to index a GPU and
make NVML calls against it that was not visiable anymore).

Now we double check if NVML can see each GPU enumerated over PCI in cases where
the nvidia module is loaded and we will eventually be performing NVML
operations on the GPUs.

Signed-off-by: Kevin Klues <kklues@nvidia.com>