Update nvidia-gpu-reset.target setup
This PR fixes an incorrect entry in the
nvidia-mig-manager.service. The "WantedBy" needs
to point to multi-user.target otherwise unless
something "Requires" or "Wants" nvidia-mig-manager
it will not ever be started.

Also the original patch that added the target did not
update the debian Makefile and RPM spec file. This
resulted in the target not being installed.

Signed-off-by: KeithMnemonic <kberger@nvidia.com>