Update rpm spec file to ensure the symlinks are kept during an upgrade
This PR updates the rpm spec file to ensure the symlinks to
configf.yaml and hooks.yaml are maintained during a package upgrade.
In addition, this also fixes an issue with the nvidia-mig-manager
service is set to disabled during an upgrade.

Currently when doing a dnf update or dnf reinstall, those two
symlinks are removed. Adding a check around these functions:

systemctl disable nvidia-mig-manager.service
systemctl daemon-reload
maybe_remove_hooks_symlink
maybe_remove_config_symlink

allows the symlinks to remain when a pacakge is reinstalled or updated
and keeps the systemd service enabled.

Closes: 154
Signed-off-by: KeithMnemonic <kberger@nvidia.com>