Update rpm spec file to ensure the symlinks are kept during an upgrade
This PR updates the rpm spec file to ensure the symlinks to
configf.yaml and hooks.yaml are maintained during a package upgrade.

Currently when doing a dnf update or dnf reinstall, those two
symlinks are removed. Adding a check around these functions:

maybe_remove_hooks_symlink
maybe_remove_config_symlink

allows the symlinks to remain when a pacakge is reinstalled or updated.

Additionally, a couple entries were added to the .gitignore as a
convenience.

Closes: 154
Signed-off-by: KeithMnemonic <kberger@nvidia.com>