Add a posttrans to ensure symlinks are kept during an upgrade
This PR adds a %posttrans section to the rpm spec file to ensure
the symlinks are kept during an upgrade. The prior fix to this
did not work correctly and the symlinks were removed.

By adding the %posttrans section, this ensures the symlinks are
recreated after the entire upgrade process is complete,
regardless of what happened during the %preun and %post phases.

Signed-off-by: KeithMnemonic <kberger@nvidia.com>